feat: rename `unlink` to `remove`
