+++ lib/element-templates/index.js
@@ -1,13 +1,13 @@
 import AppendElementTemplatesModule from './append-menu';
 import CreateElementTemplatesModule from './create-menu';
 import ReplaceElementTemplatesModule from './replace-menu';
-import UnlinkTemplatesModule from './unlink-templates';
+import RemoveTemplatesModule from './remove-templates';
 
 export default {
   __depends__: [
     AppendElementTemplatesModule,
     CreateElementTemplatesModule,
     ReplaceElementTemplatesModule,
-    UnlinkTemplatesModule
+    RemoveTemplatesModule
   ]
 };
