+++ lib/element-templates/remove-templates/RemoveTemplateReplaceProvider.js
@@ -5,7 +5,7 @@ import { getReplaceOptionGroups } from '../../util/ReplaceOptionsUtil';
  * A replace menu provider that allows to replace elements with
  * templates applied with the correspondent plain element.
  */
-export default function UnlinkTemplateReplaceProvider(popupMenu, translate, elementTemplates) {
+export default function RemoveTemplateReplaceProvider(popupMenu, translate, elementTemplates) {
 
   this._popupMenu = popupMenu;
   this._translate = translate;
@@ -14,7 +14,7 @@ export default function UnlinkTemplateReplaceProvider(popupMenu, translate, elem
   this.register();
 }
 
-UnlinkTemplateReplaceProvider.$inject = [
+RemoveTemplateReplaceProvider.$inject = [
   'popupMenu',
   'translate',
   'elementTemplates'
@@ -23,7 +23,7 @@ UnlinkTemplateReplaceProvider.$inject = [
 /**
  * Register replace menu provider in the popup menu
  */
-UnlinkTemplateReplaceProvider.prototype.register = function() {
+RemoveTemplateReplaceProvider.prototype.register = function() {
   this._popupMenu.registerProvider('bpmn-replace', this);
 };
 
@@ -33,7 +33,7 @@ UnlinkTemplateReplaceProvider.prototype.register = function() {
  *
  * @returns {Object}
  */
-UnlinkTemplateReplaceProvider.prototype.getPopupMenuEntries = function(element) {
+RemoveTemplateReplaceProvider.prototype.getPopupMenuEntries = function(element) {
 
   return (entries) => {
 
@@ -42,7 +42,7 @@ UnlinkTemplateReplaceProvider.prototype.getPopupMenuEntries = function(element) 
 
     if (this._elementTemplates && this._elementTemplates.get(element)) {
 
-      // add unlink template option
+      // add remove template option
       this.addPlainElementEntry(element, entrySet, this._translate, this._elementTemplates);
     }
 
@@ -57,12 +57,12 @@ UnlinkTemplateReplaceProvider.prototype.getPopupMenuEntries = function(element) 
 
 
 /**
- * Adds the option to replace with plain element (unlink template).
+ * Adds the option to replace with plain element (remove template).
  *
  * @param {djs.model.Base} element
  * @param {Array<Object>} entries
  */
-UnlinkTemplateReplaceProvider.prototype.addPlainElementEntry = function(element, entries, translate, elementTemplates) {
+RemoveTemplateReplaceProvider.prototype.addPlainElementEntry = function(element, entries, translate, elementTemplates) {
 
   const replaceOption = this.getPlainEntry(element, entries, translate, elementTemplates);
 
@@ -75,7 +75,7 @@ UnlinkTemplateReplaceProvider.prototype.addPlainElementEntry = function(element,
     entry
   ] = replaceOption;
 
-  // insert unlink entry
+  // insert remove entry
   entries.splice(insertIndex, 0, [ entry.id, entry ]);
 };
 
@@ -87,7 +87,7 @@ UnlinkTemplateReplaceProvider.prototype.addPlainElementEntry = function(element,
  *
  * @returns {Array<Object, number>}
  */
-UnlinkTemplateReplaceProvider.prototype.getPlainEntry = function(element, entries, translate, elementTemplates) {
+RemoveTemplateReplaceProvider.prototype.getPlainEntry = function(element, entries, translate, elementTemplates) {
 
   const {
     options,
@@ -100,7 +100,7 @@ UnlinkTemplateReplaceProvider.prototype.getPlainEntry = function(element, entrie
   }
 
   const entry = {
-    id: 'replace-unlink-element-template',
+    id: 'replace-remove-element-template',
     action: () => {
       elementTemplates.removeTemplate(element);
     },
