+++ lib/element-templates/remove-templates/index.js
@@ -1,0 +1,6 @@
+import RemoveTemplateReplaceProvider from './RemoveTemplateReplaceProvider';
+
+export default {
+  __init__: [ 'removeTemplateReplaceProvider' ],
+  removeTemplateReplaceProvider: [ 'type', RemoveTemplateReplaceProvider ]
+};
