--- lib/element-templates/unlink-templates/index.js
@@ -1,6 +1,0 @@
-import UnlinkTemplateReplaceProvider from './UnlinkTemplateReplaceProvider';
-
-export default {
-  __init__: [ 'unlinkTemplateReplaceProvider' ],
-  unlinkTemplateReplaceProvider: [ 'type', UnlinkTemplateReplaceProvider ]
-};
