+++ lib/index.js
@@ -1,3 +1,3 @@
 export { default as CreateAppendAnythingModule } from './create-append-anything';
 export { default as CreateAppendElementTemplatesModule } from './element-templates';
-export { default as UnlinkTemplatesModule } from './element-templates/unlink-templates';
+export { default as RemoveTemplatesModule } from './element-templates/remove-templates';
