+++ test/spec/element-templates/replace-menu/RemoveTemplateReplaceProvider.bpmn
