+++ test/spec/element-templates/replace-menu/RemoveTemplateReplaceProvider.element-templates.json
